requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.21.0
vllm>=0.3.0
sentencepiece>=0.1.99
protobuf>=3.20.0
